Consolisdate the units fetch to reduce the amount of work done every frame.
